Add a sanity test for local_last_updated value
In the gargoyle use case, it is possible for a process executing
_populate(reset=False) to be left with outdated local cache if another
process is executing _populate(reset=True) (e.g. value change in admin).
It is possible for _local_last_updated in the reset=False process to
be set to value greater than the value the reset=True process would
assign to cache.get(remote_cache_last_updated_key).

The new test case checks for a condition that leads to this scenario.