Update handling of CachedDict._last_checked_for_remote_changes
We want to update the value in the variable if and only if an
actual check for remote cache change took place.

Note: with this change, ModelDict can remove the request_finished
handler, as local_cache_has_expired would work as expected.
Default CachedDict timeout might need a revisit in that case.